+++ src/interlegis/portalmodelo/policy/setuphandlers.py
@@ -348,9 +348,10 @@ def setup_various(context):
 
 
 def fix_image_links_in_static_portlet(context):
-    """Fix image links in "redes-sociais" portlet. To make this independent of
-    portal site name we need to use `resolveuid/UID` as source of images
-    instead of using a fixed URL. This is called after import of portlets.xml.
+    """Fix image links in "redes-sociais" and "acesso-informacao" portlets. To
+    make this independent portal site name we need to use `resolveuid/UID` as
+    source of images instead of using a fixed URL. This is called after import
+    of portlets.xml.
     """
 
     def get_image_uid(image):
@@ -366,11 +367,19 @@ def fix_image_links_in_static_portlet(context):
     portal = api.portal.get()
     manager = getUtility(IPortletManager, name='plone.rightcolumn', context=portal)
     mapping = getMultiAdapter((portal, manager), IPortletAssignmentMapping)
-    assert 'redes-sociais' in mapping
 
+    assert 'redes-sociais' in mapping
     portlet = mapping['redes-sociais']
     images = ['ico-facebook.png', 'ico-twitter.png', 'ico-instagram.png', 'ico-youtube.png', 'ico-pinterest.png']
     for i in images:
         uid = 'resolveuid/' + get_image_uid(i)
         portlet.text = portlet.text.replace(i, uid)
     logger.debug(u'Links substituidos no portlet de redes sociais')
+
+    assert 'acesso-informacao' in mapping
+    portlet = mapping['acesso-informacao']
+    image = 'acesso-a-informacao.jpg'
+    uid = 'resolveuid/' + get_image_uid(image) + '/image_mini'
+    portlet.text = portlet.text.replace(image, uid)
+    logger.debug(u'Link substituido no portlet de acesso a informacao')
+